changing password and deleting account now works; renamed 'About' to 'Report' in Header.js; will work on Record saving and the Report page
