+++ src/ChangePassword.css
@@ -1,0 +1,12 @@
+.form-group {
+    margin-bottom: 50px;
+    position: relative;
+  }
+  
+  .error-message {
+    color: red;
+    position: absolute;
+    top: 90%;
+    left: 0;
+  }
+
