+++ src/Header.css
@@ -59,7 +59,7 @@
   color: #000;
   display: flex;
   align-items: center;
-  text-decoration: underline;
+  text-decoration: none;
 }
 
 .about-link {
