+++ src/Profile.js
@@ -2,10 +2,11 @@ import React, { useState, useEffect } from 'react';
 import './Profile.css';
 import Header from './Header';
 import { data, storage, auth } from './firebase';  // Updated import for Firestore, Storage, and Auth
-import { onAuthStateChanged, updatePassword } from 'firebase/auth';
+import { EmailAuthProvider, onAuthStateChanged, reauthenticateWithCredential, updatePassword } from 'firebase/auth';
 import { ref, uploadBytes, getDownloadURL, deleteObject } from 'firebase/storage';
-import { deleteDoc, doc, getDoc, updateDoc } from 'firebase/firestore';
-import { Button, Modal } from 'react-bootstrap';
+import { doc, getDoc, updateDoc } from 'firebase/firestore';
+import ChangePassword from './ChangePassword';
+import DeleteAccount from './DeleteAccount';
 
 const ProfilePage = () => {
   const [isEditing, setIsEditing] = useState(false);
@@ -16,13 +17,17 @@ const ProfilePage = () => {
   const [gender, setGender] = useState('');
   const [profileImage, setProfileImage] = useState('');
   const [profileImageUrl, setProfileImageUrl] = useState('');
-  const [isDeleteConfirmOpen, setIsDeleteConfirmOpen] = useState(false);
   const [username, setUsername] = useState('');
   const [isChangePasswordOpen, setIsChangePasswordOpen] = useState(false);
   const [currentPassword, setCurrentPassword] = useState('');
   const [newPassword, setNewPassword] = useState('');
   const [confirmPassword, setConfirmPassword] = useState('');
   const [changePasswordError, setChangePasswordError] = useState(null);
+
+  const [showDeleteAccount, setShowDeleteAccount] = useState(false);
+
+  const handleCloseDeleteAccount = () => setShowDeleteAccount(false);
+  const handleShowDeleteAccount = () => setShowDeleteAccount(true);
 
   useEffect(() => {
     //const user = auth.currentUser;
@@ -134,28 +139,6 @@ const ProfilePage = () => {
     }
   };
 
-  const handleDeleteAccount = async ()=> {
-    const user = auth.currentUser;
-    if (user) {
-      try {
-        await deleteDoc(doc(data, 'users', user.uid));
-
-        const profileImageRef = ref(storage, 'profileImages/' + user.uid);
-        await deleteObject(profileImageRef).catch((error) => {
-          if (error.code !== 'storage/object-not-found') {
-            throw error;
-          }
-        });
-
-        await user.delete();
-        window.location.href = '/landing';
-      } catch (e) {
-        console.error('Error deleting account:', e);
-        alert('Failed to delete account. Please try again.');
-      }
-    }
-  }
-
   const openChangePassword = () => {
     setIsChangePasswordOpen(true);
     setChangePasswordError(null);
@@ -168,125 +151,6 @@ const ProfilePage = () => {
     setConfirmPassword('');
     setChangePasswordError(null);
   };
-
-  const openDeleteConfirm = () => {
-    setIsDeleteConfirmOpen(true);
-  };
-  
-  const closeDeleteConfirm = () => {
-    setIsDeleteConfirmOpen(false);
-  };
-
-  const handleChangePassword = async ()=> {
-    if (newPassword !== confirmPassword) {
-      alert('New password and confirm password do not match.');
-      return;
-    }
-
-    try {
-      const user = auth.currentUser;
-      if (user) {
-        await updatePassword(user, newPassword);
-        setCurrentPassword('');
-        setNewPassword('');
-        setConfirmPassword('');
-        setChangePasswordError(null);
-        alert('Password updated successfully.');
-        setIsChangePasswordOpen(false);
-      }
-    } catch (e) {
-      console.error('Error updating password: ', e);
-      alert('Failed to update password. Please try again.')
-    }
-  }
-
-  const DeleteConfirmPopup = ({ show, onClose, onDeleteConfirm }) => {
-  <div>
-    <Modal 
-      show={isDeleteConfirmOpen} 
-      onHide={onClose}>
-      <Modal.Header closeButton style={{ display: 'flex', flexDirection: 'column-reverse' }}>
-        <Modal.Title>Confirm Delete Account</Modal.Title>
-      </Modal.Header>
-      <Modal.Body>
-        Are you sure you want to delete your account? This action cannot be undone.
-      </Modal.Body>
-      <Modal.Footer>
-        <Button variant="secondary" onClick={onClose}>
-          Cancel
-        </Button>
-        <Button variant="danger" onClick={onDeleteConfirm}>
-          Delete Account
-        </Button>
-      </Modal.Footer>
-    </Modal>
-  </div>
-};
-
-  const ChangePasswordModal = ({
-    currentPassword,
-    setCurrentPassword,
-    newPassword,
-    setNewPassword,
-    confirmPassword,
-    setConfirmPassword,
-    handleChangePassword,
-    error,
-    onClose, // Added onClose prop
-  }) => (
-    <div className="change-password-modal">
-        <Modal 
-          show={true} 
-          onHide={onClose}
-          backdrop='static'
-          keyboard={false}
-          centered
-          scrollable
-          >
-        <Modal.Header closeButton style={{ display: 'flex', flexDirection: 'column-reverse' }}>
-          <Modal.Title>Change Password</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          <div className="form-group">
-            <label htmlFor="currentPassword">Current Password</label>
-            <input
-              type="password"
-              id="currentPassword"
-              value={currentPassword}
-              onChange={(e) => setCurrentPassword(e.target.value)}
-            />
-          </div>
-          <div className="form-group">
-            <label htmlFor="newPassword">New Password</label>
-            <input
-              type="password"
-              id="newPassword"
-              value={newPassword}
-              onChange={(e) => setNewPassword(e.target.value)}
-            />
-          </div>
-          <div className="form-group">
-            <label htmlFor="confirmPassword">Confirm Password</label>
-            <input
-              type="password"
-              id="confirmPassword"
-              value={confirmPassword}
-              onChange={(e) => setConfirmPassword(e.target.value)}
-            />
-          </div>
-          {error && <p className="error-message">{error}</p>}
-        </Modal.Body>
-        <Modal.Footer>
-          <Button variant="secondary" onClick={onClose}>
-            Close
-          </Button>
-          <Button variant="primary" onClick={handleChangePassword}>
-            Save Changes
-          </Button>
-        </Modal.Footer>
-      </Modal>
-    </div>
-  );
 
   return (
     <div>
@@ -325,7 +189,7 @@ const ProfilePage = () => {
               phone={phone}
               setPhone={setPhone}
             />
-            <SecuritySection isEditing={isEditing} openDeleteConfirm={openDeleteConfirm} openChangePassword={openChangePassword}  />
+            <SecuritySection isEditing={isEditing} openDeleteConfirm={handleShowDeleteAccount} openChangePassword={openChangePassword}  />
             <RecordsSection />
           </section>
           {/* {isEditing && (
@@ -336,26 +200,9 @@ const ProfilePage = () => {
             </div>
           )} */}
         </div>
+        <ChangePassword isOpen={isChangePasswordOpen} onClose={closeChangePassword}/>
+        <DeleteAccount show={showDeleteAccount} handleClose={handleCloseDeleteAccount}/>
       </main>
-      {isDeleteConfirmOpen && (
-      <DeleteConfirmPopup
-        onDeleteConfirm={handleDeleteAccount}
-        onClose={closeDeleteConfirm}
-      />
-      )}
-      {isChangePasswordOpen && (
-        <ChangePasswordModal
-          currentPassword={currentPassword}
-          setCurrentPassword={setCurrentPassword}
-          newPassword={newPassword}
-          setNewPassword={setNewPassword}
-          confirmPassword={confirmPassword}
-          setConfirmPassword={setConfirmPassword}
-          handleChangePassword={handleChangePassword}
-          error={changePasswordError}
-          onClose={closeChangePassword}
-        />
-      )}
     </div>
   );
 };
@@ -493,31 +340,7 @@ const ContactInfo = ({ isEditing, email, setEmail, phone, setPhone }) => (
 const SecuritySection = ({ isEditing, openDeleteConfirm, openChangePassword }) => (
   <section className="security-wrapper">
     <h2 className="section-title">Security</h2>
-    {/* <div className="form-group">
-      <h2 className="section-names">Password</h2>
-      <input 
-      id="password" 
-      type="password" 
-      value="test" 
-      readOnly={!isEditing} 
-      style={{color: 'gray'}} />
-    </div> */}
     <div className="form-group">
-      {/* <h2 className="section-names">Two-Factor Authentication</h2>
-      <div className="toggle-switch">
-        <input
-          type="checkbox"
-          id="twoFactorAuth"
-          name="twoFactorAuth"
-          className="toggle-switch-checkbox"
-          checked
-          disabled
-        />
-        <label className="toggle-switch-label" htmlFor="twoFactorAuth">
-          <span className="toggle-switch-inner"></span>
-          <span className="toggle-switch-switch"></span>
-        </label>
-      </div> */}
       <button className='change-password' onClick={openChangePassword}>Change Password</button>
       <div className='button-deleteAccount'>
         <button className='deleteAccount' onClick={openDeleteConfirm}>Delete Account</button>
